Remove unused models + separate LL from Modular behaviors
