+++ configs/default/experiment.yaml
@@ -1,0 +1,63 @@
+global_exp : &global
+    cuda: True
+
+lifelong_exp: &lifelong_exp
+#    log_dir: /local/veniat/lileb/ray_results
+#    log_dir: /net/blackorpheus/veniat/lileb/ray_results
+    log_dir: /data/veniat/lileb/ray_results
+    smoke_test: false
+
+    name: ll
+    cuda: true
+    use_ray: true
+    use_ray_logging: False
+    plot_all: False
+    redis_address: 'localhost:6381'
+    use_processes: True
+    local_mode: False
+    n_it_max: 10000
+    n_ep_max: null
+    schedule_mode: null
+    augment_data: False
+    normalize: False
+    single_pass: False
+    grace_period: 500
+    num_hp_samplings: 1
+
+    val_per_task: True
+    stream_setting: False
+
+    split_optims: True
+
+    n_tasks: 50
+
+    patience: 100
+#    visdom_traces_folder: /local/veniat/lileb/visdom_traces
+#    visdom_traces_folder: /net/blackorpheus/veniat/lileb/visdom_traces
+    visdom_traces_folder: /data/veniat/lileb/visdom_traces
+    batch_sizes: [512, 1024]  # Train and Eval batch sizes
+    norm_models: []
+#    norm_models: [0-0-0-inde, inde-upper-bound]
+
+    log_steps: [1, 2, 3, 4, 5]
+    lca: 5
+    log_epoch: False
+
+#    task_save_folder: /local/veniat/lileb/datasets
+#    task_save_folder: /net/blackorpheus/veniat/lileb/datasets
+    task_save_folder: /data/veniat/lileb/datasets
+    load_tasks_from:
+
+    ref_params_id:
+
+    plot_tasks: False
+
+
+stream_tuning:
+    <<: *lifelong_exp
+
+task_tuning:
+    <<: *lifelong_exp
+
+plot_archs:
+    <<: *lifelong_exp
