+++ configs/default/optimizers.yaml
@@ -1,0 +1,12 @@
+optim_commons: &commons
+    weight_decay: 0
+
+adam:
+    <<: *commons
+    lr: 1.e-3
+    betas: [0.9, 0.999]
+
+sgd:
+    <<: *commons
+    lr: 1.e-1
+    momentum: .9
