+++ configs/streams/s_in.yaml
@@ -27,13 +27,13 @@ datasets:
 
 modules:
     ll_models:
-#        inde-batch:
-#            _name: change-layer-resnet
+        inde-batch:
+            _name: change-layer-resnet
 #            split_last: False
-#            share_layer: [0, 0, 0, 0, 0 ,0]
-#            ray_resources:
-#                cpu: 2
-#                gpu: 0.3
+            share_layer: [0, 0, 0, 0, 0 ,0]
+            ray_resources:
+                cpu: 2
+                gpu: 0.3
 #        finetune:
 #            _name: change-layer-resnet
 #            split_last: False
@@ -260,22 +260,23 @@ modules:
 #                cpu: .2
 #                gpu: 0.25
 ##
-#        PSSN-search-6-fw:
-#            _name: pssn-resnet
-#            strat: search_all
-#            n_source_models: 1
-#            connections: [fw, bw]
-#            grid_params:
-#                optim:
-#                    - lr:  [1.e-2, 1.e-3]
-#                      weight_decay: [0, 1.e-4, 1.e-5]
-#                learner-params:
-#                    arch_loss_coef: 0
-#                    entropy_coef: 0
-#                    split_training: True
-#            ray_resources:
-#                cpu: 2
-#                gpu: 1
+        PSSN-search-6-fw:
+            _name: pssn-resnet
+            strat: search_all
+            block_id: [0, 1, 1, 2, 2, 3 ,3, 4, 5, 6]
+            n_source_models: 1
+            connections: [fw, bw]
+            grid_params:
+                optim:
+                    - lr:  [1.e-2, 1.e-3]
+                      weight_decay: [0, 1.e-4, 1.e-5]
+                learner-params:
+                    arch_loss_coef: 0
+                    entropy_coef: 0
+                    split_training: True
+            ray_resources:
+                cpu: 2
+                gpu: 1
 #        PSSN-Alex-search-6-fw:
 #            _name: pssn-alexnet
 #            strat: search_all
@@ -317,23 +318,23 @@ modules:
 #                cpu: 2
 #                gpu: 0.3
 #
-        wide-hat:
-            _name: hat
-            wide: 2.5
-            #            hidden_size: [1]
-            #            channel_scaling: False
-            grid_params:
-                optim:
-                    lr:  [5.e-2, 1.e-2, 5.e-3, 1.e-3]
-                    weight_decay: [0, 1.e-4, 1.e-5]
-                #                    lr:  [1.e-3]
-                #                    weight_decay: [0]
-                learner-params:
-                    lamda: [0.75]
-                    smax: [400]
-            ray_resources:
-                cpu: 2
-                gpu: 0.5
+#        wide-hat:
+#            _name: hat
+#            wide: 2.5
+#            #            hidden_size: [1]
+#            #            channel_scaling: False
+#            grid_params:
+#                optim:
+#                    lr:  [5.e-2, 1.e-2, 5.e-3, 1.e-3]
+#                    weight_decay: [0, 1.e-4, 1.e-5]
+#                #                    lr:  [1.e-3]
+#                #                    weight_decay: [0]
+#                learner-params:
+#                    lamda: [0.75]
+#                    smax: [400]
+#            ray_resources:
+#                cpu: 2
+#                gpu: 0.5
 
 optimizers:
     optim:
