+++ src/models/__init__.py
@@ -5,17 +5,13 @@
 
 from collections import OrderedDict
 
-from src.modules.PNN_llmodel import PNN_LLmodel
-from src.modules.PSSN_llmodel import ProgressiveSSN
-from src.modules.SuperNet_llmodel import SuperNetLLModel
-from src.modules.change_layer_llmodel import ChangeLayerLLModel
-from src.modules.ewc_llmodel import EWCLLModel
-from src.modules.experience_replay_llmodel import ExperienceReplayLLModel
-from src.modules.fine_tune_head_llmodel import FineTuneHeadLLModel
-from src.modules.fine_tune_leg_llmodel import FineTuneLegLLModel
-from src.modules.multitask_head_llmodel import MultitaskHeadLLModel
-from src.modules.multitask_leg_llmodel import MultitaskLegLLModel
+
 from . import _utils
+from .PNN_llmodel import PNN_LLmodel
+from .PSSN_llmodel import MNTDP
+from .change_layer_llmodel import ChangeLayerLLModel
+from .ewc_llmodel import EWCLLModel
+from .experience_replay_llmodel import ExperienceReplayLLModel
 from .hat_llmodel import HATLLModel
 from .zoo_llmodel import ZooLLModel
 
@@ -23,14 +19,6 @@ from .zoo_llmodel import ZooLLModel
 def get_module_by_name(name):
     if name.startswith('change-layer'):
         return ChangeLayerLLModel
-    if name == 'multitask-head':
-        return MultitaskHeadLLModel
-    if name == 'multitask-leg':
-        return MultitaskLegLLModel
-    if name == 'finetune-head':
-        return FineTuneHeadLLModel
-    if name == 'finetune-leg':
-        return FineTuneLegLLModel
     if name.startswith('ewc'):
         return EWCLLModel
     if name.startswith('er'):
@@ -39,10 +27,8 @@ def get_module_by_name(name):
         return HATLLModel
     if name == 'pnn':
         return PNN_LLmodel
-    if name == 'ssn':
-        return SuperNetLLModel
     if name.startswith('pssn'):
-        return ProgressiveSSN
+        return MNTDP
     if name == 'zoo':
         return ZooLLModel
     if name.endswith('-dict'):
