+++ src/utils/plotting.py
@@ -26,10 +26,6 @@ from pandas import DataFrame, MultiIndex
 from plotly.express import parallel_coordinates
 from sklearn.metrics import auc
 
-from constants import MODEL_ALIASES, MODELS, FILE_NAMES, BASE_RESNET_FLOPS, \
-    BATCH_SIZE, BASE_RESNET_PARAMS, M_AVG_ACC, M_FORGETTING, M_COMPUTE, \
-    M_MEMORY, M_LCA, M_T_OUT, OUTER_CROSS_VAL, M_T_IN, INCH, BASE_ALEX_FLOPS, \
-    RADAR_NAMES, RADAR_ALIASES, BARPLOT_IDS, CNN, LONG_TRAJ_KEEPERS, M_BW
 from src.utils.misc import get_env_url, paretize_exp, count_params
 
 logger = logging.getLogger(__name__)
@@ -973,736 +969,3 @@ def process_final_results(main_vis, res_dict, exp_name, visdom_conf,
             first_plot = False
     return global_summary
 
-
-# MODEL_ALIASES = defaultdict(lambda x: x)
-
-# MODELS = sorted(MODELS)
-# raise ValueError(MODELS)
-
-
-def _get_steps(best_traj, ):
-    steps = best_traj['duration_best_it'].sum()
-    if '/env/2527_Trial' in best_traj['env_url'][0] \
-            and 'PSSN-search-6-fw' in best_traj['env_url'][0]:
-        assert steps == 2570, steps
-        steps = 7 * steps
-    if '/env/2492_Trial' in best_traj['env_url'][0] \
-            and 'PSSN-search-6-fw' in best_traj['env_url'][0]:
-        assert steps == 30530, steps
-        steps = 4 * steps
-    return steps
-
-
-def _get_macs(best_traj):
-    env_string = best_traj['env_url'][0]
-    print(env_string)
-    if 'ewc-online' in env_string:
-        print('EWC-O')
-        steps_first = best_traj['duration_best_it'][0]
-        macs = steps_first * BASE_RESNET_FLOPS * BATCH_SIZE
-
-        remaining_steps = sum(best_traj['duration_best_it'][1:])
-        assert remaining_steps + steps_first == _get_steps(best_traj)
-        step_macs = BATCH_SIZE * BASE_RESNET_FLOPS + 2 * BASE_RESNET_PARAMS
-        macs += remaining_steps * step_macs
-    elif 'ewc-full' in env_string:
-        print('EWC-F')
-        macs = 0
-        for i, iters in enumerate(best_traj['duration_best_it']):
-            step_macs = BATCH_SIZE * BASE_RESNET_FLOPS \
-                        + i * 2 * BASE_RESNET_PARAMS
-            macs += iters * step_macs
-    elif 'PNN' in env_string:
-        print('PNN')
-        macs = 0
-        for i, iters in enumerate(best_traj['duration_best_it']):
-            macs += iters * (2 * i + 1) * BASE_RESNET_FLOPS * BATCH_SIZE
-    elif 'er-ring' in env_string or 'er-reservoir' in env_string:
-        print('ER')
-        steps = _get_steps(best_traj)
-        macs = steps * BASE_RESNET_FLOPS * 2 * BATCH_SIZE
-    elif 'wider-hat' in env_string:
-        steps = _get_steps(best_traj)
-        macs = steps * BASE_ALEX_FLOPS[6.5] * BATCH_SIZE
-    elif 'wide-hat' in env_string:
-        steps = _get_steps(best_traj)
-        macs = steps * BASE_ALEX_FLOPS[3.2 if CNN else 4] * BATCH_SIZE
-    elif 'hat' in env_string or 'Alex' in env_string:
-        steps = _get_steps(best_traj)
-        macs = steps * BASE_ALEX_FLOPS[1] * BATCH_SIZE
-    else:
-        print('Other')
-        steps = _get_steps(best_traj)
-        macs = steps * BASE_RESNET_FLOPS * BATCH_SIZE
-    return macs
-
-
-def _get_results(best_traj, tasks, mod_summary):
-    tasks, metrics = tasks
-    res = {}
-    if tasks:
-        t1 = tasks[0]
-        first_t = t1 + 1
-        acc_t1 = best_traj[f'Test_T{t1}'][t1]
-        res[f'Acc $T_{first_t}$'] = np.round(acc_t1, 2)
-        for i, t2 in enumerate(tasks[1:], 1):
-            t_alias = 'T_{}{}'.format(first_t, "'" * i)
-            acc_t2 = best_traj[f'Test_T{t2}'][t2]
-            res[f'Acc ${t_alias}$'] = np.round(acc_t2, 2)
-            diff = np.round(acc_t2 - acc_t1, 2)
-            res[f'$\Delta_{{T_{first_t}, {t_alias}}}$'] = diff
-    if M_BW in metrics:
-        l = len(best_traj['Test_T0'])
-        res[M_BW] = best_traj['Test_T0'][l-1] - best_traj['Test_T0'][0]
-        # res[M_BW+'2'] = best_traj['Test_T0'][-1] - best_traj['Test_T0'][0]
-
-    for met in metrics:
-        if met == M_AVG_ACC:
-            res[M_AVG_ACC] = np.round(best_traj['avg_acc_test'].iloc[-1],
-                                          2)
-        elif met == M_FORGETTING:
-            acc_t = mod_summary['Acc t']
-            res[met] = np.round(res[M_AVG_ACC] - acc_t, 2)
-
-        # elif met == 'Steps':
-        #     steps = _get_steps(best_traj)
-        #     res[met] = np.round(steps / 1.e3, 1)
-        elif met == M_COMPUTE:
-            res[met] = int(np.round(2*_get_macs(best_traj)/1.e12))
-        elif met == M_MEMORY:
-            params = best_traj['total_params'].iloc[-1]
-            mem = np.round(4*params/1.e6, 1)
-            res[met] = mem
-        elif met == M_LCA:
-            res[met] = np.round(best_traj['lca'].mean(), 2)
-        else:
-            logger.warning(f'Unknown metric {met}')
-    return res
-
-
-def fill_result(best_traj, tasks, table, mod_summary):
-    res = _get_results(best_traj, tasks, mod_summary)
-    for k, v in res.items():
-        table[k].append(v)
-
-
-def fill_ph(tasks, table, ph=float('nan')):
-    tasks, metrics = tasks
-    if tasks:
-        first_t = tasks[0] + 1
-        table[f'Acc $T_{first_t}$'].append(ph)
-        for i, t2 in enumerate(tasks[1:], 1):
-            t_alias = 'T_{}{}'.format(first_t, "'" * i)
-            table[f'Acc ${t_alias}$'].append(ph)
-            table[f'$\Delta_{{T_{first_t}, {t_alias}}}$'].append(ph)
-    for met in metrics:
-        table[met].append(ph)
-
-
-def normalize_table(table, ref, columns):
-    if ref not in table['Model']:
-        logger.warning(f'Trying to normalize by {ref}, but not present in table !')
-        return
-    idx = table['Model'].index(ref)
-    for prop in columns:
-        column = table[prop]
-        ref = column[idx]
-        if ref is float('nan'):
-            logger.warning(f'No value for column {prop}')
-            continue
-        table[prop] = [f'{n} ({(n/ref)}x)' for n in column]
-
-
-def plot_delta_table(tasks, res, global_summary, vis):
-    all_models = _remove_ducplicates(MODELS + list(res.keys()))
-
-    table = defaultdict(list)
-    for ll_name in all_models:
-        table['Model'].append(MODEL_ALIASES[ll_name] if ll_name in MODEL_ALIASES else ll_name)
-        if ll_name in res:
-            best_traj, _ = res.get(ll_name)
-            mod_summary = global_summary.loc[ll_name]
-            fill_result(best_traj, tasks, table, mod_summary)
-        else:
-            fill_ph(tasks, table)
-
-    df = pandas.DataFrame(table).set_index('Model')
-    # df = df.sort_values('Model')
-    html = df.to_html(classes=['table', 'table-bordered', 'table-hover'],
-                      na_rep='-')
-    latex = df.to_latex(index=True, bold_rows=True, escape=False, na_rep='-')
-    vis.text(html)
-    vis.text(f'<pre>{latex}</pre>')
-    print(latex)
-    return df
-
-
-def get_agg_results(results_df, std=True, string=True):
-    strats = set(map(lambda x: f'{x["strat"]} {x["n_tasks"]}',
-                     results_df.values()))
-    if len(strats) == 0:
-        logger.warning('No results to show')
-        return
-    elif len(strats) > 1:
-        logger.warning("Got results from differrent strats, can't " \
-                             "aggregate that {}".format(strats))
-        return
-
-    cat = pd.concat(map(itemgetter('df'), results_df.values()),
-                    keys=results_df.keys())
-    exp_ids = defaultdict(list)
-    for k, v in results_df.items():
-        exp_ids[v['seed']].append(k)
-    # exp_ids = set(cat.index.get_level_values(0))
-
-    groups = cat.groupby(by='Model')
-    # std = False
-    if std and string:
-        res = groups.agg(aggreg_row_std_str)
-    elif std:
-        res = groups.agg([aggreg_row_mean, aggreg_row_er])
-    else:
-        res = groups.agg(aggreg_row_simple)
-        # res = groups.agg(['mean', 'std'])
-
-    return exp_ids, res, strats.pop()
-    # print(res)
-    # res = res.groupby(level=[0, 1]).apply(
-    #     lambda x: x.round(3).astype(str).apply('±'.join, 0))
-    # print(res)
-    # for k, v in results_df.items():
-    #     print(k, v)
-
-
-def aggreg_row_simple(row, ref='independent', columns=['Steps (K)']):
-    if row.dropna().empty:
-        return float('nan')
-    else:
-        return f'{np.mean(row.dropna().values).round(3)}'
-
-def aggreg_row_mean(row, ref='independent', columns=['Steps (K)']):
-    if row.dropna().empty:
-        return float('nan')
-    else:
-        return np.mean(row.dropna().values)
-
-
-def aggreg_row_er(row):
-    if row.dropna().empty:
-        return float('nan')
-    else:
-        return np.std(row.dropna().values) / len(row.dropna())
-
-def aggreg_row_std_str(row):
-    if row.dropna().empty:
-        return float('nan')
-    else:
-        return f'{np.mean(row.dropna().values).round(2)} ' \
-               f'$\pm$ {(np.std(row.dropna().values) / len(row.dropna())).round(2)} '\
-               # f'({len(row.dropna())})'
-
-
-def _remove_ducplicates(sequence):
-    seen = set()
-    return [x for x in sequence if not (x in seen or seen.add(x))]
-
-
-def normalize_df(df, offset=.2, inv=None):
-    if inv is None:
-        inv = []
-    for col in inv:
-        if col in df:
-            df[col] = -df[col]
-    df -= df.min()
-    df /= df.max()
-    df = df * (1-offset) + offset
-    return df
-
-
-def radar_chart(full_res, multi=True, rows=None, columns=None):
-    if rows is None:
-        rows = full_res.index
-    plot_df = full_res[columns].loc[rows]
-    # plot_df.columns = plot_df.columns.droplevel(0)
-    plot_df = normalize_df(plot_df, inv=[M_COMPUTE, M_MEMORY])
-    # plot_df = plot_df.transpose().reset_index()
-
-    plot_df = plot_df.fillna(value=1)
-    if multi:
-        res = radar_chart_matplotlib_multi(plot_df)
-    else:
-        res = radar_chart_matplotlib_simple(plot_df)
-    res.show()
-    return res
-
-def radar_char_column_names(columns):
-    categories = []
-    for col in columns:
-        # if col == M_MEMORY:
-        #     col = f'         $-${M_MEMORY}'
-        # elif col == M_COMPUTE:
-        #     col = f'            $-${M_COMPUTE}'
-        # elif col == M_FORGETTING:
-        #     col = f'        $-${M_FORGETTING}'
-        # elif col == M_T_IN:
-        #     col = f'{M_T_IN}        '
-        # elif col == M_T_OUT:
-        #     col = f'{M_T_OUT}       '
-        col = RADAR_NAMES[col]
-        categories.append(col)
-    categories = [cat.replace('\mbox{', '\\;') for cat in categories]
-    return categories
-
-def radar_chart_matplotlib_simple(df, size=23):
-    # ------- PART 1: Create background
-    # plt.style.use('seaborn')
-    # number of variable
-    # categories = [col.replace('\mbox{', '\\;') for col in df.columns]
-    categories = radar_char_column_names(df.columns)
-    # categories = list(df.columns)
-    N = len(categories)
-
-
-    # What will be the angle of each axis in the plot? (we divide the plot / number of variable)
-    angles = [n / float(N) * 2 * pi for n in range(N)]
-    angles += angles[:1]
-
-    # Initialise the spider plot
-    ax = plt.subplot(111, polar=True)
-
-    # If you want the first axis to be on top:
-    ax.set_theta_offset(pi / 2)
-    ax.set_theta_direction(-1)
-
-    # Draw one axe per variable + add labels labels yet
-    plt.xticks(angles[:-1], categories, size=size)
-
-    # Draw ylabels
-    ax.set_rlabel_position(0)
-    plt.yticks([.2, .4, .6, .8, 1], [".2", ".4", ".6", ".8", "1"],
-               color="grey", size=size*0.8, visible=False)
-    plt.ylim(0, 1)
-
-    # ------- PART 2: Add plots
-
-    # Plot each individual = each line of the data
-    # I don't do a loop, because plotting more than 3 groups makes the chart unreadable
-
-    # Ind1
-    my_palette = plt.cm.get_cmap("Set2", 6)#len(df.index)+1)
-    line_styles = ['-', '--', '-.', ':', (-2.5, (7, 2)), (0, (1, 4))]
-    for i, (name, trace) in enumerate(df.iterrows()):
-        # ax = plt.subplot(i,2,2, polar=True)
-        values = trace.values.flatten().tolist()
-        values += values[:1]
-        w = 2
-        if name.startswith('MNTDP'):
-            w = 2.5
-        c = my_palette(i)
-        # print(c)
-        # input()
-        # if name=='HAT (Alexnet)':
-        #     c = 'blue'
-        #     i=3
-        # if name == 'HAT (Wide Alexnet)':
-        #     i=4
-        #     c = my_palette(i)
-        if name in RADAR_ALIASES:
-            name = RADAR_ALIASES[name]
-        ax.plot(angles, values, color=c, linewidth=w,
-                linestyle=line_styles[i], label=name)
-        ax.fill(angles, values, color=c, alpha=0.1)
-
-    # # Ind2
-    # values = df.loc[1].drop('group').values.flatten().tolist()
-    # values += values[:1]
-    # ax.plot(angles, values, linewidth=1, linestyle='solid', label="group B")
-    # ax.fill(angles, values, 'r', alpha=0.1)
-
-    # Add legend
-    plt.legend(loc='lower left',
-               # bbox_to_anchor=(1.5, 0.1),
-               # bbox_to_anchor=(1.15, 0.65),
-               bbox_to_anchor=(-0.7, 0.8),
-               prop={'size': size})
-    # plt.gcf().set_size_inches(9, 5)
-    plt.gcf().set_size_inches(12, 7)
-    plt.tight_layout(1)
-    # plt.legend()
-    plt.show()
-    return plt
-
-
-def radar_chart_matplotlib_multi(df):
-    categories = radar_char_column_names(df.columns)
-    # categories = list(df.columns)
-    N = len(categories)
-    plt.cla()
-    plt.clf()
-    # What will be the angle of each axis in the plot? (we divide the plot / number of variable)
-    angles = [n / float(N) * 2 * pi for n in range(N)]
-    angles += angles[:1]
-
-    num_traces = len(df.index)
-    n_cols = 3
-    n_rows = np.ceil(num_traces/n_cols)
-
-    def make_spider(idx, title, trace, color):
-        # Initialise the spider plot
-        ax = plt.subplot(n_rows, n_cols, idx+1, polar=True, )
-
-        # If you want the first axis to be on top:
-        ax.set_theta_offset(pi / 2)
-        ax.set_theta_direction(-1)
-
-        # Draw one axe per variable + add labels labels yet
-        if idx == 0:
-            plt.xticks(angles[:-1], categories, color='grey', size=12)
-        else:
-            plt.xticks(angles[:-1], ['']*len(categories), color='grey', size=1)
-
-        # Draw ylabels
-        ax.set_rlabel_position(0)
-        plt.yticks([.2, .4, .6, .8, 1], [".2", ".4", ".6", ".8", "1"],
-                   color="grey", size=7, visible=False)
-        plt.ylim(0, 1)
-
-        # Ind1
-        values = trace.values.flatten().tolist()
-        values += values[:1]
-        ax.plot(angles, values, color=color, linewidth=2, linestyle='solid')
-        ax.fill(angles, values, color=color, alpha=0.4)
-
-        # Add a title
-        plt.title(title, size=11, color=color, y=1.15)
-
-    # ------- PART 2: Apply to all individuals
-    # initialize the figure
-    # my_dpi = 96
-    # plt.figure(figsize=(1000 / my_dpi, 1000 / my_dpi), dpi=my_dpi)
-    plt.gcf().set_size_inches(9, 12)
-
-    # Create a color palette:
-    my_palette = plt.cm.get_cmap("Set2", len(df.index))
-
-    # Loop to plot
-    for idx, (name, trace) in enumerate(df.iterrows()):
-        make_spider(idx=idx, title=name, trace=trace,
-                    color=my_palette(idx))
-
-    # plt.gca().get_legend()
-    plt.tight_layout()
-    return plt
-
-
-def do_the_barplots(df, strat, path, keepers=None, legend=False, cols=4):
-    return
-    # MNIST
-    if keepers is None:
-        keepers = df.index
-    n_cols = 3
-    if isinstance(df.columns, MultiIndex):
-        with_error = True
-        df_err = df.xs('aggreg_row_er', level=1, axis=1)
-        plot_df_err = df_err[df_err.columns[-n_cols:]]
-        plot_df_err = plot_df_err.loc[keepers].transpose()
-        df = df.xs('aggreg_row_mean', level=1, axis=1)
-        print(f'Got errors: {plot_df_err}')
-    else:
-        with_error = False
-    plot_df = df[df.columns[-n_cols:]]
-    plot_df = plot_df.loc[keepers].transpose()
-    # idx = list(plot_df.index)
-    # idx[2], idx[3] = idx[3], idx[2]
-    # plot_df = plot_df.reindex(idx)
-    # colors = plt.cm.get_cmap("tab20", len(plot_df.columns))
-    # n_styles = len(plot_df.columns)
-    n_styles = 7
-    colors = plt.cm.get_cmap("tab20", n_styles)
-    colors = [colors(i) for i in range(n_styles)]
-    patterns = ["||", "\\\\", "//", "+", "---", ".", "*", "x", "o", "O",
-                '...', '.-', '+++', '/.', '.|']
-    fig, axes = plt.subplots(nrows=1, ncols=n_cols)
-    # fig, axes = plt.subplots(nrows=2, ncols=2)
-
-    print(plot_df.index)
-
-    if with_error:
-        items = zip(plot_df.iterrows(), plot_df_err.iterrows())
-    else:
-        items = plot_df.iterrows()
-
-    handles = None
-    for i, data in enumerate(items):
-        if with_error:
-            ((metric_v, row_v), (metric_err, row_err)) = data
-        else:
-            metric_v, row_v = data
-            metric_err = metric_v
-            row_err = [None] * len(row_v)
-        assert metric_v == metric_err
-        metric = metric_v
-        # ax = axes[int(i/2), i % 2]
-        ax = axes[i]
-        new_handles = []
-        for j, ((model, itm), err) in enumerate(zip(row_v.iteritems(), row_err)):
-            if model in BARPLOT_IDS:
-                style_id = BARPLOT_IDS[model]
-            else:
-                style_id = j
-            ax.bar(
-                x=j,
-                height=itm,
-                color=colors[style_id],
-                width=0.9,
-                yerr=err,
-                # hatch=patterns[j]
-            )
-
-            ax.bar(x=j,
-                   height=itm,
-                   color='None',
-                   alpha=0.5,
-                   edgecolor='black',
-                   hatch=patterns[style_id%len(patterns)])
-
-            rect = patches.Rectangle((0, 0), 1, 1, color=colors[style_id])
-            rect.set_hatch(patterns[style_id % len(patterns)])
-            rect.set_edgecolor('k')
-            new_handles.append(rect)
-
-        if handles is None:
-            handles = new_handles
-
-        # row.plot.bar(
-        #     title=metric,
-        #     colors=colors,
-        #     ax=ax,
-        #     width=0.9,
-        #     hatch=patterns[i])
-        ax.set_xlabel('')
-        ax.set_ylabel('')
-        # ax.set_ylabel(metric)
-        if metric == M_MEMORY:
-            ax.set_yscale('log')
-            ax.set_title(f'Log {metric}$\\downarrow$')
-        elif metric == M_FORGETTING:
-            ax.invert_yaxis()
-            ax.set_title(metric + '$\\downarrow$')
-        elif metric == M_AVG_ACC:
-            margin = 0.05 if CNN else 0.01
-            ax.set_ylim([row_v.min() - margin, min(row_v.max() + margin, 1)])
-            # ax.set_ylim([0.5, 1])
-            ax.set_title(metric + '$\\uparrow$')
-        else:
-            ax.set_title(metric + '$\\downarrow$')
-
-
-        # ax.set_x_tick(None)
-        ax.tick_params(
-            axis='x',  # changes apply to the x-axis
-            which='both',  # both major and minor ticks are affected
-            bottom=False,  # ticks along the bottom edge are off
-            top=False,  # ticks along the top edge are off
-            labelbottom=False)  # labels along the bottom edge are off
-        ax.tick_params(
-            axis='y',  # changes apply to the x-axis
-            # nbins=4,
-            # style='sci',
-            which='both',  # both major and minor ticks are affected
-            rotation=90)  # labels along the bottom edge are off
-        ax.set_axisbelow(True)
-        for label in ax.get_yticklabels():
-            label.set_verticalalignment('center')
-        if metric != M_MEMORY:
-            ax.locator_params(nbins=3, axis='y')
-            ax.ticklabel_format(axis='y', style='sci')
-        ax.yaxis.grid(color='gray', linestyle='dashed')
-    # bar = plot_df.plot.bar(color=colors, subplots=True)
-    # fig = bar.get_figure()
-    # handles = [patches.Rectangle((0, 0), 1, 1, color=colors[i],)
-    #            for i in range(n_styles)]
-    for i, h in enumerate(handles):
-        h.set_hatch(patterns[i%len(patterns)])
-        h.set_edgecolor('k')
-    # edgecolor = 'black',
-
-
-    # hatch = patterns[i])
-
-    if legend:
-        # 1x4 legend
-        if len(plot_df.columns) <= 11:
-            plt.legend(handles, list(plot_df.columns),
-                       loc='lower left',
-                       ncol=1,
-                       bbox_to_anchor=(1, 0),
-                       prop={'size': 13})
-        else:
-            plt.legend(handles, list(plot_df.columns),
-                       loc='right',
-                       ncol=1,
-                       bbox_to_anchor=(2.4, 0.55),
-                       prop={'size': 8})
-
-    for ax in axes:
-        for item in ([ax.title, ax.xaxis.label, ax.yaxis.label] +
-                     ax.get_xticklabels() + ax.get_yticklabels()):
-            item.set_fontsize(15)
-            # ax.set_yticklabels(va='bottom')
-            # ax.tick_params(axis="y", va='bottom')
-    # 4x4 legend
-    # plt.legend(handles, list(plot_df.columns),
-    #     loc='upper left', bbox_to_anchor=(1.3, 3), prop={'size': 7})
-
-    mult = 1 if legend else 0.8
-    fig.set_size_inches(8.5*mult, 2.5)
-    fig.tight_layout()
-    f_name = FILE_NAMES[strat][:-4]
-
-    fig.savefig(path/(f_name + '_barplot.svg'))
-    fig.savefig(path/(f_name + '_barplot.png'))
-    fig.savefig(path/(f_name + '_barplot.pdf'))
-
-    fig.show()
-    # input()
-    # raise ValueError
-
-    print(plot_df)
-
-def plot_traces(x, y, y_err, y_name, traces_names, ax, labels=True):
-    # fig, ax = plt.subfigures(21, figsize=(8, 5))
-    my_palette = plt.cm.get_cmap("Set2", len(traces_names))
-    line_styles = ['-', '--', '-.', (1, (5, 2, 18, 2)), (-2.5, (8, 1)), ':']
-    print(len(traces_names))
-    print(len(traces_names))
-    print(len(traces_names))
-    print(traces_names)
-    for i, (name, y, y_err) in enumerate(zip(traces_names, y, y_err)):
-        ax.plot(x, y, color=my_palette(i), linestyle=line_styles[i%len(line_styles)],
-                label=name if labels else None, linewidth=2)
-        ax.fill_between(x, y - y_err, y + y_err, alpha=0.3,
-                        facecolor=my_palette(i))
-
-    ax.set(ylabel=y_name)
-    # plt.plot(x, y1,
-    #          color='red',
-    #          linewidth=1.0,
-    #          linestyle='--'
-    #          )
-
-
-def plot_traj_curves(data_dict, out_path):
-    res_dict = {}
-    for idx, data in data_dict.items():
-        for mod, res in data['res'].items():
-            assert (idx, mod) not in res_dict
-            res_dict[(idx, mod)] = res[0]
-
-    print(res_dict.keys())
-    out_cross_val = defaultdict(list)
-    for k in list(res_dict.keys()):
-        for pref in OUTER_CROSS_VAL:
-            if k[1].startswith(pref):
-                out_cross_val[(k[0], pref)].append((k, res_dict.pop(k)))
-                break
-
-    print(res_dict.keys())
-    print(f'OUTER CROSSVAL : {out_cross_val.keys()}')
-    for k, candidates in out_cross_val.items():
-        best = None
-        best_val = None
-        for name, res in candidates:
-            avg_val = res['avg_acc_val_so_far'][-1]
-            if best is None or avg_val > best_val:
-                best = name
-                best_val = avg_val
-                best_res = res
-        print(f'BEST: {best}')
-        best_k = '-'.join(best[1].split('-')[:-1])
-        print(f'BEST k: {best_k}')
-        res_dict[(best[0], f'best-{best_k}')] = best_res
-
-    print(res_dict.keys())
-
-    df = pd.DataFrame(res_dict, )
-    keep = ['avg_acc_test_so_far', 'lca', 'total_params', 't']
-    df = df.loc[keep]
-    df2 = df
-    fig, axs = plt.subplots(3,1, sharex=True)
-    x = df.loc['t'].tolist()[0]
-
-
-    # df2 = df.groupby(level=1, axis=1).agg(aggregate_trajectories)
-    df = df.transpose()
-    df = df.groupby(level=1).agg([ #aggregate_trajectories)
-        mean_s,
-        se_s,
-    ])
-
-    df = df.loc[LONG_TRAJ_KEEPERS]
-    # input(df)
-    trace_names = [MODEL_ALIASES[n] for n in df.index]
-
-    avg_acc = [np.array(itm) for itm in df['avg_acc_test_so_far', 'mean_s'].values]
-    avg_acc_err = [np.array(itm) for itm in df['avg_acc_test_so_far', 'se_s'].values]
-    # lca = [np.array(itm) for itm in df.loc['lca'].values]
-    n_params = [np.array(itm) * 4 for itm in df['total_params', 'mean_s'].values]
-    n_params_err = [np.array(itm) * 4 for itm in df['total_params', 'se_s'].values]
-    plot_traces(x, avg_acc, avg_acc_err, M_AVG_ACC, trace_names, axs[2])
-    plot_traces(x, avg_acc, avg_acc_err, M_AVG_ACC, trace_names, axs[1])
-    plot_traces(x, n_params, n_params_err, M_MEMORY, trace_names, axs[0], labels=False)
-    axs[2].set(xlabel='Task id')
-    lgd = fig.legend(
-        loc='bottom left',
-        # bbox_to_anchor=(.532, .923),
-        bbox_to_anchor=(0.63, 1.06),
-        fontsize=12,
-        ncol=int(np.ceil(len(trace_names) / 3)),
-    )
-    for ax in axs:
-        for item in ([ax.title, ax.xaxis.label, ax.yaxis.label] +
-                     ax.get_xticklabels() + ax.get_yticklabels()):
-            item.set_fontsize(11)
-    save_p = {
-        'bbox_extra_artists': (lgd,),
-        'bbox_inches': 'tight'
-    }
-
-    fig.subplots_adjust(hspace=0.0)
-    # fig.set_size_inches(3*6.5/INCH, 3*3.2/INCH)
-    fig.set_size_inches(3*7/INCH, 3*4/INCH)
-    fig.savefig(out_path/'long-stream-traj.svg', **save_p)
-    fig.savefig(out_path/'long-stream-traj.pdf', **save_p)
-    fig.savefig(out_path/'long-stream-traj.png', **save_p)
-
-    fig.show()
-    return fig
-
-
-def aggregate_trajectories(traj, method='mean'):
-    return traj.apply(partial(mean_std_series, method=method), axis=1)
-
-
-def se_s(s):
-    return mean_std_series(s, 'se')
-
-def std_s(s):
-    return mean_std_series(s, 'std')
-
-
-def mean_s(s):
-    return mean_std_series(s, 'mean')
-
-
-def mean_std_series(series, method):
-    vals = np.array(series.values.tolist())
-    if method == 'mean':
-        return vals.mean(axis=0).tolist()
-    elif method == 'std':
-        return vals.std(axis=0).tolist()
-    elif method == 'se':
-        se = vals.std(axis=0) / vals.shape[0]
-        return se.tolist()
-    else:
-        raise ValueError(f'Unknown method {method}')
